Restructure files to use cobra in main.go
